added abstract
-split components into classes with inheritance
