+++ src/view/task-list.js
@@ -1,27 +1,11 @@
-import {createElement} from "../utils.js";
+import AbstractView from "./abstract.js";
 
 const createTaskListTemplate = () => {
   return `<div class="board__tasks"></div>`;
 };
 
-export default class TaskList {
-  constructor() {
-    this._element = null;
-  }
-
+export default class TaskList extends AbstractView {
   get template() {
     return createTaskListTemplate();
-  }
-
-  get element() {
-    if (!this._element) {
-      this._element = createElement(this.template);
-    }
-
-    return this._element;
-  }
-
-  removeElement() {
-    this._element = null;
   }
 }
